Fix issue where the ID3D11ShaderResourceView* was converted to a uint32_t, so ImGui crashed when trying to render it as a texture

+++ Kerberos/src/Kerberos/Renderer/Framebuffer.h
@@ -58,7 +58,7 @@ namespace Kerberos
 
 		virtual void Resize(uint32_t width, uint32_t height) = 0;
 
-		virtual uint32_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
+		virtual uint64_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
 
 		virtual FramebufferSpecification& GetSpecification() = 0;
 		virtual const FramebufferSpecification& GetSpecification() const = 0;
