+++ Kerberos/src/Platform/OpenGL/OpenGLTextureCube.h
@@ -14,7 +14,7 @@ namespace Kerberos
 
 		void Bind(uint32_t slot = 0) const override;
 
-		uint32_t GetRendererID() const override { return m_RendererID; }
+		uint64_t GetRendererID() const override { return m_RendererID; }
 		const std::string& GetName() const override { return m_Name; }
 		uint32_t GetWidth() const override;
 		uint32_t GetHeight() const override;
